Fix definition of email type in node
The included mongoose version may be different from the one included by
an application. This results in a definition of schema types on the
internal mongoose module.

Fix this by explicitly letting the user inject a mongoose module which
has also the advantage of dropping a dependency. As this breaks
backwards compatibility, bump the major version.

Tested with mongoose 3.9.7 and 4.0.3.